tcMacro: enforce atomicity of types
In the macro typechecker language we have 'IntLike' & 'FloatLike', which
behave like a family of types in Haskell-land.

To avoid generating Haskell type signatures such as:

  forall i. IntLike i -> IntLike i

we must consider instances in which the argument to 'IntLike'/'FloatLike'
remains a metavariable to be a family of instances, and behave the same
as if there were multiple matching instances.

This commit also:

  - Removes HsBindgen.Runtime.Syntax, as it only defined the Haskell
    IntLike and FloatLike type synonyms which were meant to work around
    the above issue.
  - Adds an assertion at the end of tcMacro to ensure it only produces
    atomic types.