Add --version option
A `--version` option is added that prints version information for both
`hs-bindgen` and `libclang`.  The raw Clang version string is printed,
to give users as much information as possible.  This can be helpful to
confirm which version of `libclang` is being used by `hs-bindgen` in an
environment with more than one version available.

Note that the `libclang` version depends on the `libclang` shared
library that is dynamically linked.