Introduce an `IncompleteArray` type to `hs-bindgen-runtime`
This new type functions similarly to the `ConstantArray` type. Where
`ConstantArray` represents a C array of known size, `IncompleteArray` represents
a C array of unknown size. The goal is to use `IncompleteArray` in generated
bindings where possible for user convenience, but also to emphasise that `Ptr
(IncompleteArray a)` points to an array of `a` as a whole, whereas the
previously used `Ptr a` could mean either that it points to single element of
type `a`, or to a sequence of elements of type `a`.