Refactor HashIncludeArg
We need to construct `HashIncludeArg`s even when there are trace
messages.  Function `hashIncludeArgEither` is refactored to
`hashIncludeArg`, which returns a tuple (including any number of
`HashIncludeArgMsg`.  Users may want to use the "unsafe" constructor as
well, when a constant is known to be valid, so `HashIncludeArg(..)` is
exported.  Note that the TH API still uses `FilePath` and emits the
trace messages; use of `HashIncludeArg(..)` is for test and library
usage.

The `HashIncludeArgMsg` traces are changed to default to `Notice`.

The `Aeson` instances are removed.  Backslashes or absolute paths should
not result in failure; they should just emit trace messages.  Both
binding specifications and command-line parsing does the parsing using
pure functions.

For binding specifications, `hashIncludeArg` is used to accumulate the
trace messages, and a new `BindingSpecReadMsg` constructor is used to
return them along with the rest of the trace messages.

For command-line parsing, inputs are parsed to
`type UncheckedHashIncludeArg = FilePath`, and a `checkInputs` helper
uses `hashIncludeArgWithTrace` to emit any trace messages and construct
the `HashIncludeArg`s.