Add Comment pretty printer capability to Qu monad
There's some technical limitations that makes it not possible to pretty
print generated Haddocks for TH in Golden tests. These limitations are:

- the fact that we TH `Dec` AST does not contain documentation information
- to generate `Dec` all we can use is `Guasi` interface which is not
  capable of injecting/attaching documentation. For that we'd need
  `Quasi`.

With this being said this commit makes changes to the `Qu` Monad so that
it keeps track of all the `Name` -> `Comment` mappings so that we can
then output them in the right place. This is easy to do for top level
declarations, however if we want to pretty print the comments at the
record fields level we have to traverse the `Dec` AST.

The implemented TH pretty printer is faithful to the original one with the exceptions of the following:

- It inserts comments in the right places
- In some places it is faithful to older versions of TH pretty printer
  to make it compatible with our CI test matrix