Remove commentOrigin from TH golden tests
Disabled Hs and PP golden tests for doxygen_docs.h file.

In unnamed structures libclang can not get the original C name of the
declaration and so it points to the file path, line number and column of
where the struct is defined. This is troublesome for golden tests
because it uses absolute paths.