Fix elaborated type issue
In the `va_list` fix (#978), the code expects the underlying type of the
`typedef` to be elaborated.  This is only the case in later versions of
Clang!  With this commit, the kind of the underlying type is checked and
`clang_Type_getNamedType` is only called when the type is elaborated.