Add a TODO and accompanying test for macro "newtypes"
These macro "newtypes" redefine an existing identifier as a macro, but the macro
simply desugars to the pre-exesting identifier. Such cases exis in the wild, for
example in `stdio.h`:

```c
typedef int FILE;

extern FILE *const stdin;
extern FILE *const stdout;
extern FILE *const stderr;

\#define stdin  (stdin)
\#define stdout (stdout)
\#define stderr (stderr)
```

The current behaviour of `hs-bindgen` is to not generate a binding and emit a
"redeclaration" trace message. Should `hs-bindgen` instead allow this
corner-case of a redeclared identifier?