Refactor `Declaration` to `DeclarationClassification`
The data type and the accompanying `classifyDeclaration` function had
struct-/union-/enum-specific terminology, but the classification works for other
types of declarations as well, such as variable and function declarations. We
refactor `Declaration` to use slightly more general terminology that applies to
a larger class of declarations, and we expand its documentation.