Factour out an `addressStubDecs` function in the `Hs` backend pass
Currently we're generating a stub C function for each global (constant) variable
that returns the address of said variable. We would like to start generating
these stubs for addresses of other declarations as well, such as functions. For
this reason, we factor out a dedicated `addressStubDecs` function that deals
with generating a stub C function that returns the address of any declaration
that it is given.