Mangle the C names of address stubs
Resolves #946

Previously, we were just prepending `get_` and appending `_ptr` to the original
C name, and using that as the name of the address stub C function. However, we
have to make C names unique or we risk linker errors, so now we used mangled,
unique C names for the address stubs instead. The original stub names are added
as C comments to the stub definitions.