Generate address stubs for functions
This makes it easier for users of generated bindings to pass C function pointers
to other C functions through the Haskell FFI, as long as those C functions are
declared in a header file. Haskell functions still have to be converted to
function pointers manually if they are to be passed to C functions.