Refactor getMainHeadersAndInclude
Our `getMainHeader` implementation used `panicPure` on error.  There
were errors when processing builtin declarations, but we never evaluated
them.

I refactored to `getMainHeadersAndInclude` and ensured that it was never
used on builtin declarations, but the implementation continued to use
`panicPure`.

This commit refactors `getMainHeadersAndInclude` to return an `Either`
type, so that we can use `panicIO` and avoid such silent issues.

There was *another* issue!  We get the `DeclInfo` even for include
directives, and they also caused an error that was never evaluated.
This commit resolves that error as well.