Remove PrelimDeclId from matchParse
We used to filter out builtin macros using `matchParse`, but now we want
to also use a `ParsePredicate` to select which parts of an include graph
to show.  Handling of builtin macros needs to be handled specially
anyway (#1087).

This change removes the builtin check from `matchParse` and implements
it in `foldDecl`.  A `TODO` comment referring to #1087 is added.