Use artifacts
This currently does not work well because it runs *all* boot and
frontend actions, resulting in many more traces and (possibility of)
errors.  This will be resolved when we refactor artifacts to only
execute actions that are necessary (with caching).

An unfortunate side-effect is that the command now has *all*
`hs-bindgen` options, even though many are irrelevant.  With the current
design, we just have to accept that as a cost of being able to run the
same code path.

The `frontendIncludeGraph` artifact now includes an include graph
predicate (used to filter which vertices are shown) as well as the
include graph.  The predicate is needed to display the include graph,
but we still have the full include graph available for other purposes.

The `writeIncludeGraph` artifact is refactored to output to either
`STDOUT` or a file.