Add hs-bindgen version to binding specifications
Syntax:

```yaml
version:
    hs_bindgen: 0.1.0
    binding_specification: '1.0'
```

Note that the `hs_bindgen` version number does *not* have to be quoted,
since it has three components and therefore cannot be parsed as a
number.  The `binding_specification` version number *does* have to be
quoted, since it has two components and therefore can be parsed as a
number.