Rename to CommentRef
The comment reference type was previously called just `Reference`.  This
name is too general.  This commit renames the type to `CommentRef` to
make it clear that it is a comment reference, not one of the other
reference types that we use.  Abbreviation `Ref` is consistent with the
other reference types.