Move default generated module name
It was defined in a `Default` instance for `ModuleName` in
`HsBindgen.Language.Haskell`.  This module is for general Haskell, and
the `ModuleName` type is used for more than just specifying the names of
generated modules.  The default belongs in
`HsBindgen.Backend.HsModule.Translation`, but a complication is that the
`ConfigPP` uses `ModuleName` directly instead of `HsModuleOpts`.  I
resolved this by defining and exporting `defModuleName`, which is used
in the `Default` instance for both `HsModuleOpts` and `ConfigPP`.