Reimplement getMainHeadersAndInclude
This was implemented using `DynGraph.findTargets`, which I implemented
to get exactly what is needed but was a bit out-of-place because it is
not a generally-useful graph function.  I replaced it with
`DynGraph.findEdges`, which is a bit more general.  It finds edges from
the specified starting vertex and edges to terminal vertices in paths
from that starting vertex.  Taking advantage of the structure of the
graph, where all paths terminate at the root header, it can get the same
information with a much simpler implementation.