Implement omitted types in binding spec generation
We already tracked omitted types in `ResolveBindingSpecs`, in a
`Set C.QualName`.  This commit changes the type to
`Map C.QualName SourcePath`, because the source path is needed to
determine the headers.  `resolveBindingSpecs` returns this map along
with the translation unit and trace messages.

We already implemented a `GetMainHeadersAndInclude` function, which is
used to get header information for Haddock documentation.  This commit
adds `GetMainHeaders`, which just gets the main headers, implemented
using the existing function.

For now, both the omitted type map and the `GetMainHeaders` function are
added as `Frontend` artifacts.  After we move instance resolution into
the frontend, we will be able to refactor this so that binding
specification generation is done within the frontend, which would allow
us to instead provide a generated binding specification artifact.

Binding specification generation is refactored to use `GetMainHeaders`
to determine headers for all types, including the omitted types.  A
list of main headers is no longer passed in.

Golden test `binding_spec_simple` is added, using a prescriptive binding
specification.  A new `testPrescriptiveBindingSpec` field is added to
`TestCase` to support this.