Only generate high-level wrappers for const-qualified arrays
Only if the array elements are const qualified do we know for sure that the
array is read-only. In such cases, we generate a high-level wrapper with a pure
Haskell array value as input. If the arrays can be witten to, then we don't use
a pure Haskell array value but a pointer to the first array element instead.