Delay macro translation
In the translation from the C AST to our Haskell AST, we now keep the
`CheckedMacroExpr` as-is. Previously we were translatig this to some
intermediate representation, and then from there to `SExpr`. In this commit we
_still_ do that, but we do it later, when we translate from `Hs` to `SHs`. The
next step will be to fuse these two translation steps, so that we eliminate the
intermediate types entirely.

However, even as it stands, this is already an improvement over the status quo,
because `SigmaType` and co and now all internal to a new
`HsBindgen.Backend.SHs.Macro` module.