Add module mismatch error
It is an error if a prescriptive binding specification is for a
different module than the one being processed.  When using the CLI, the
module specified via a CLI option must match.  When using TH, the module
executing the Template Haskell must match.

The internal representation of a binding specification is *not*
module-specific.  We therefore need to use a flag to ensure that only
one module mismatch error is output as we process the declarations.