Make binding specs module-specific
The internal `BindingSpec` type is refactored to be specific to a
Haskell module, matching that aspect of the JSON/YAML (Aeson)
representation.  This is done so that we do not lose semantics.

For prescriptive binding specifications, this is straightforward.

For external binding specifications, there is no empty/zero value.  In
addition, merging external binding specifications so that we can resolve
them with one `Map` lookup per declaration/use is an important
optimization.  New type `MergedBindingSpecs` is introduced to resolve
both of these.  We decided to make this type a
`Map C.QualName [(Set SourcePath, ResolvedBindingSpec)]`.  This requires
*two* lookups per declaration/use, which we consider acceptable.

The private `load` function is removed, since prescriptive binding
specifications are now loaded differently than external binding
specifications.  The separate load functions are implemented directly in
the public because the `stdlib` configuration is defined there.

Lookup functions now also return the Haskell module.