Nix: Resolve intricacies when overriding the version of LLVM used
The dependency graph involving LLVM versions is complex. We want to override the
used version of LLVM in some places, but not all. (If we completely change the
LLVM version, even the Rust compiler itself is recompiled).

This commit carefully overrides the version of LLVM used in places where we need
the specific LLVM version, but not in places were the default LLVM version is
alright, avoiding unnecessary rebuilds.

Also update the Nix Flake lock file.