Add export list to preprocess mode.
This export list follows an invariant that each module exports all their
definitions and types as well as re-exports underlying newtype
constructors. This is needed to workaround the FFI compilation issue
where GHC requires all constructors in scope.

NOTE: This almost solves the problem mentioned above. There's one corner
      case we didn't consider which is the fact that this is possible:

```hs
newtype A = A { un_A :: Other.Module.A }
```

RPM does this, in fact... which means we can't really export the two A's
in the same module :(