Refactor Target
The `TargetEnv` type is removed, and significant environments are now
distinguished in `Target`.  `Target` is therefore now a sum type that
enumerates the complete set of supported platforms.

Function `targetTriple` associates each `Target` with a canonical target
triple string.  Function `parseTargetTriple` parses these strings, only
accepting the exact canonical target triples.  Users *must* use
canonical target triple strings in command-line arguments and binding
specifications.

Function `parseTargetTripleLenient` should only be used to parse the
host target triple as reported by LLVM/Clang, translating it to one of
the canonical target triples.  Target triples are a mess, so this parser
is *not* expected to be perfect.  In cases where the reported target
triple cannot be parsed, users should configure one of the canonical
target triples.  Unit tests are added for this function, including all
of the target triples that I have observed on Windows and macOS.