Use a dynamic matrix in the 'build and test' GHA workflow
This makes the logs nicer to read because steps are not folded together like
with composite actions. As an added benefit, we are now also able to simplify
the required checks to list only the "Check success" job. That job only succeeds
if all preceding jobs, the actual builds and tests, succeed.