Add target boot artifact
We have been treating targets as something that only concerns
cross-compilation.  The (host) target may have a significant effect on
how bindings are generated, however, even when not doing
cross-compilation.

This commit adds a target boot artifact, which means that we always
determine the current target.  This changes a number of things, but
perhaps the most significant is that we now check that we are generating
bindings for a supported target, something that we did not check before
when using the host target.

The Clang arguments and current target are determined together.  First,
the Clang arguments are determined using the `ClangConfig`.  If the
target is specified, then those Clang arguments and the target are used.
If the target is not specified, however, those Clang arguments are used
to call `libclang` and get the current target triple, and the lenient
parser is used to attempt to translate the current target triple to a
supported target.  If it can be translated to a supported target, we
then reconstruct Clang arguments using that target.  Subsequent calls
therefore use one of our canonical target triples.

This implementation throws an `UnableToDetermineTargetException`
exception if the call to `libclang` fails or the target triple returned
by `libclang` does not translate to a supported target.  This is done so
that we do not continue with a default/inappropriate or unsupported
target.

When a target is not specified and the target triple from `libclang` is
*not* successfully translated to a supported target, an `Error` message
is traced that shows the target triple string.  It includes instructions
to "configure a supported target to resolve this error."

When a target is not specified and the target triple from `libclang` is
successfully translated to a supported target, an `Info` message is
traced that shows the parsed target as well as the target triple string.