Introduce `Translation.UniqueSymbol`
This intentionally does not yet change _how_ we produce the symbol, so that the
fixtures remain unchanged. However, the responsibility of including the
`Safety` or not is moved to the caller, as this does not make sense in all
cases; doing it this way will result in a more readable `UniqueSymbol.source`,
once we start including that in the generated output (next commit).