Move `BaseName` into the hash
For C names this makes no difference, the result should still be unique,
provided that the hash is strong enough. For Haskell names it's irrelevant,
since these generated names are anyway in the scope of a module and not
intended for export (on the import side we always import qualified, so there is
no danger of duplicate symbol errors even _if_ we would get hash collisions).

The benefit of this approach is that we can then in the next commit define a
version `locallyUnique` alongside `globallyUnique`, which can be used to
generate names that only need to be unique within a Haskell module.