Implement SupportedInstances
This commit implements a new `HsBindgen.Instances` module without
changing existing code, which will be done in a separate commit.

`TypeClass` is copied from `HsBindgen.Language.Haskell`, and additional
type classes are added.  This is not done in
`HsBindgen.Language.Haskell` because many type classes are specific to
`hs-bindgen` and are therefore not standard.

`Strategy` is copied from `HsBindgen.BindingSpec.Private.V1`.  Defining
it alongside `TypeClass` makes sense.

`SupportedInstances` and `SupportedStrategies` are refactored from my
old branch, and the `Default` instance of `SupportedInstances` defines
the instances and strategies that are supported by `hs-bindgen`.

If we decide to allow users to configure this, to change which instances
are generated/derived by default for example, then we should probably
provide an API that ensures that a subset of the supported API is
configured.

Command `hs-bindgen-cli info instances` is added.  It prints out the
instance information in the (Markdown) format that I have been pasting
into the issues.  We can easily support multiple output formats if
desired, of course.