Use `ConstPtr` for result types passed through pointers
We had switched to using `ConstPtr` in the function arguments of bindings, but
we were not properly handling yet the case where result types (like structs and
unions) are communicated through a pointer. We add a new `allocaAndPeekConst`
function and we use this instead of `allocaAndPeek` whenever we are dealing with
a pointer to a const struct or union.