Refactor C standard
We decided to always get the C standard from `libclang`.  Users may
specify a Clang standard via Clang arguments.

Clang standards specify *both* the C standard and use of GNU extensions.
New type `ClangCStandard` is the product of `CStandard` and `Gnu`.
While these types should generally be specified together, but it is
convenient to case match them separately.  See the `libclang` commit for
details.

This commit removes `cStandard` and `gnu` from `ClangArgsConfig`.
`Boot` is refactored to get the C standard using `libclang`.  Various
refactoring is done to work with these changes.

The `--standard` and `--gnu` options are removed from the CLI.

The C standard types are removed from the `TH` interface.  New module
`HsBindgen.Clang.CStandard` re-exports `Clang.CStandard`, so we can make
the module public or re-export types in the `TH` interface if desired,
but it is no longer configuration.

The `c-expr-runtime` test is changed to use C17, since the Clang
argument for it is not dependent on the LLVM/Clang version.

The `hs-bindgen` tests use `c23` or `c2x` depending on the LLVM/Clang
version.  This will soon be refactored!