Fix use of private API in DelayedIO
Function `genBindingSpec` is changed to take a `Format` argument and
return a `ByteString`.  Function `genBindingSpecYaml`, used for tests,
is removed.

`DelayedIO` is refactored to not deal with binding specifications
directly.  `BindingSpecContent` is removed, `ByteStringContent` is
added, and `TestContent` is renamed to `StringContent` to reflect the
actual type.

`HsBindgen.writeBindingSpec` now renders the `ByteString`, using the
public API only, and uses `ByteStringContent`.