Order ctypes by source location in generated specs
We can only order `ctypes` by source location when source location
information is available.  We have this information when generating a
binding specification, but not in general.

A function to order IDs (`C.DeclId -> C.DeclId -> Ordering`) is added
to the *private* `encode` function.  Function `defCompareCDeclId`
compares using the rendered name, and the *public* `encode` function
uses it.  If this function returns `Eq`, the headers are used as a
tie-breaker.

Function `genBindingSpec` now takes `IncludeGraph` and `DeclIndex`
arguments, and it uses these to implement a comparison function that
orders by source location.

To implement this, omitted and squashed entries in the `DeclIndex`
needed to be given source locations, not just source paths.  This
involves some minimal refactoring of the `ResolveBindingSpecs` and
`MangleNames` passes.